% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleMutations.R
\name{assembleMutations}
\alias{assembleMutations}
\title{extract clonal mutations from a Canopy run and its VAFs}
\usage{
assembleMutations(output_tree, VAF, genelist = NULL)
}
\arguments{
\item{output_tree}{canopy output}

\item{VAF}{variant allele frequencies}

\item{genelist}{coordinates for the genes (default is to autopopulate)}
}
\value{
a data.frame
}
\description{
In order to plot Canopy configs nicely, we use timescape.
This function massages Canopy output into mutation data.
Note: if genelist has impacts, coords, breakpoints, etc, they are added!
}
