% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClonalPrev.R
\name{getClonalPrev}
\alias{getClonalPrev}
\title{extract clonal prevalence from a Canopy run for plotting}
\usage{
getClonalPrev(output_tree, tp_order)
}
\arguments{
\item{output_tree}{Canopy output}

\item{tp_order}{timepoint ordering}
}
\value{
\preformatted{           a data.frame 
}
}
\description{
In order to plot Canopy configs nicely, we use timescape.
This function massages the Canopy output for timescape.
}
